Moved component into correct directory
